9338 : CreatedBy & ModifiedBy columns in external schema tables are no longer displaying user name, FK to core.Users
SVN r12853 |2010-01-28 23:06:15 +0000

+++ src/org/labkey/test/bvt/DbUserSchemaTest.java
@@ -326,7 +326,7 @@ public class DbUserSchemaTest extends BaseSeleniumWebTest
         for (int i = 0; i < pks.length; i++)
         {
             Map<String, Object> row = resp.getRows().get(i);
-            Long rowid = (Long)row.get("RowId");
+            Integer rowid = (Integer)row.get("RowId");
             assertEquals("Expected requested rowid and selected rowid to be the same", rowid.intValue(), pks[i]);
             
             String text = (String)row.get("Text");
